move and finish
