+++ skpro/regression/linear/_sklearn.py
@@ -1,4 +1,4 @@
-"""Adapters to sklearnn linear regressors with probabilistic components."""
+"""Adapters to sklearn linear regressors with probabilistic components."""
 # copyright: skpro developers, BSD-3-Clause License (see LICENSE file)
 # based on sktime pipelines
 
@@ -359,131 +359,5 @@ class BayesianRidge(_DelegateWithFittedParamForwarding):
             "lambda_2": 2e-6,
             "compute_score": True,
             "fit_intercept": False,
-        }
-        return [param1, param2]
-
-
-class PoissonRegressor(_DelegateWithFittedParamForwarding):
-    """Poisson regression, direct adapter to sklearn PoissonRegressor.
-
-    Generalized Linear Model with a Poisson distribution.
-    This regressor uses the 'log' link function.
-
-    Parameters
-    ----------
-    alpha : float, default=1.0
-        Constant that multiplies the penalty term. Defaults to 1.0.
-        See the notes for the exact mathematical meaning of this
-        parameter. alpha = 0 is equivalent to unpenalized GLMs.
-
-    fit_intercept : bool, default=True
-        Whether to fit an intercept term.
-    
-    solver : {'lbfgs', 'newton-cholesky'}, default='lbfgs'
-        Algorithm to use in the optimization problem.
-
-    'lbfgs' is an optimization algorithm that approximates the BFGS algorithm
-    'newton-cholesky' uses a Newton-CG variant of Newton's method.
-        
-    max_iter : int, default=100
-        The maximal number of iterations for the solver.
-
-    tol : float, default=1e-4
-        The convergence tolerance. If it is not None, training will stop
-        when (loss > best_loss - tol) for n_iter_no_change consecutive
-        epochs.
-
-    verbose : int, default=0
-        For the 'sag' and 'lbfgs' solvers set verbose to any positive
-        number for verbosity.
-
-    warm_start : bool, default=False
-        When set to True, reuse the solution of the previous call to fit as
-        initialization, otherwise, just erase the previous solution.
-
-    Attributes
-    ----------
-    coef_ : array-like of shape (n_features,)
-        Coefficients of the regression model (mean of distribution)
-
-    intercept_ : float
-        Independent term in decision function.
-
-    n_iter_ : int
-        The actual number of iterations before reaching the stopping criterion.
-
-    n_features_in_ : int
-        Number of features seen during :term:'fit'.
-
-    feature_names_in_ : ndarray of shape (n_features,)
-        Names of features seen during :term:'fit'.
-    """
-
-    def __init__(
-        self,
-        alpha=1.0,
-        fit_intercept=True, 
-        max_iter=100, 
-        tol=1e-4, 
-        verbose=0, 
-        warm_start=False
-    ):
-        self.alpha = alpha
-        self.fit_intercept = fit_intercept
-        self.max_iter = max_iter
-        self.tol = tol
-        self.verbose = verbose
-        self.warm_start = warm_start
-
-        from sklearn.linear_model import PoissonRegressor
-
-        skl_estimator = PoissonRegressor(
-            alpha=alpha, 
-            fit_intercept=fit_intercept, 
-            max_iter=max_iter, 
-            tol=tol, 
-            verbose=verbose, 
-            warm_start=warm_start
-        )
-
-        skpro_est = SklearnProbaReg(skl_estimator)
-        self._estimator = skpro_est.clone()
-
-        super().__init__()
-
-    FITTED_PARAMS_TO_FORWARD = [
-        "coef_", 
-        "intercept_", 
-        "n_iter_", 
-        "n_features_in_", 
-        "feature_names_in_"
-    ]
-
-    @classmethod
-    def get_test_params(cls, parameter_set="default"):
-        """Return testing parameter settings for the estimator.
-
-        Parameters
-        ----------
-        parameter_set : str, default="default"
-            Name of the set of test parameters to return, for use in tests. If no
-            special parameters are defined for a value, will return `"default"` set.
-
-        Returns
-        -------
-        params : dict or list of dict, default = {}
-            Parameters to create testing instances of the class
-            Each dict are parameters to construct an "interesting" test instance, i.e.,
-            `MyClass(**params)` or `MyClass(**params[i])` creates a valid test instance.
-            `create_test_instance` uses the first (or only) dictionary in `params`
-        """
-        param1 = {}
-        param2 = {
-            "alpha": 2.0,
-            "fit_intercept": False,
-            "max_iter": 200,
-            "tol": 2e-4,
-            "verbose": 1,
-            "warm_start": True
         }
         return [param1, param2]
